requirements.txt
APScheduler==3.6.3
asgiref==3.2.9
Django==3.0.7
django-apscheduler==0.3.0
djangorestframework==3.11.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
